refactor devices, decouple BaseDevice from Manipulator
