+++ packages/data-sdk/src/devices/IRealtimeSubscriber.ts
@@ -1,0 +1,9 @@
+import { RealtimeDataStream, RealtimeListener } from "./device.types";
+
+export interface IRealtimeSubscriber {
+  addRealtimeListener(listener: RealtimeListener): void;
+  removeRealtimeListener(listener: RealtimeListener): void;
+
+  startListeningToRealtimeDataStream(stream: RealtimeDataStream): void;
+  stopListeningToRealtimeDataStream(stream: RealtimeDataStream): void;
+}
